# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agora API
  version: 0.1.0
  summary: Ankara Üniversitesi öğrenci bilgi sistemi ve öğrenme ortamı API'si
  description: |
    Agora API'nin sözleşmesidir (contract-first): uç noktalar önce burada tasarlanır,
    backend bu sözleşmeye göre yazılır, web istemcisi de buradan üretilir.

    **Hata yanıtları** RFC 9457 `application/problem+json` biçimindedir. İstemciler
    `code` alanına göre karar verir. `detail` kullanıcıya gösterilebilir ama
    değişebilir.

    **Kimlik doğrulama:** korumalı uç noktalar `Authorization: Bearer <access_token>`
    ister. Access token 15 dakika geçerlidir. Süresi dolunca `POST /api/v1/auth/refresh`
    ile yenisi alınır.

    **Yetkilendirme:** yetkiler token'a gömülmez, her korumalı istekte sunucuda güncel
    olarak çözülür. Rol değişikliği ya da hesabın askıya alınması anında etkilidir.
    Belirli bir yetki gerektiren uçlarda yetki `x-agora-permission` alanında yazar.

    **Parola değiştirme zorunluluğu:** `must_change_password` açık bir kullanıcı sadece
    kendi hesabıyla ilgili uçlara (`/api/v1/me...`) erişebilir. Diğer bütün korumalı
    uçlar `403 PASSWORD_CHANGE_REQUIRED` döner. İstemci kullanıcıyı parola değiştirme
    ekranına yönlendirmelidir.

    **İki adımlı doğrulama (MFA):** hesabında MFA açık kullanıcının girişi iki adımlıdır:
    `POST /api/v1/auth/login` parola doğruysa token yerine `mfa_required: true` ve kısa
...
    görünür.

...
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/users:
    get:
      tags: [Kullanıcılar]
      summary: Hesapları listeler
      description: Kullanıcı adına göre sıralı, cursor ile sayfalanır.
      operationId: listUsers
      x-agora-permission: user:read
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          description: Kullanıcı adı, e-posta ya da ad soyadda geçen metin
          schema: { type: string, maxLength: 100 }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/UserStatus" }
        - name: kind
          in: query
          schema: { type: string, enum: [STUDENT, STAFF] }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Hesaplar
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/UserSummary" }
                  next_cursor: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Kullanıcılar]
      summary: Hesap oluşturur
      description: |
        Kişi, öğrenci ya da personel kaydı ve hesap birlikte oluşturulur. Hesap
        `PENDING` başlar ve kullanıcıya 72 saat geçerli bir aktivasyon e-postası
        gönderilir: kullanıcı kendi parolasını belirleyince hesap etkinleşir. Yönetici
        parola görmez ya da belirlemez. Numara kullanıcı adı olur.
      operationId: createUser
      x-agora-permission: user:manage
      x-agora-mfa: true
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: Hesap oluşturuldu
          headers:
            Location:
              schema: { type: string, examples: ["/api/v1/users/01a11b7f-..."] }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: "`ACCOUNT_EXISTS`: numara ya da e-posta başka bir hesapta kayıtlı"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/users/{id}/status:
    put:
      tags: [Kullanıcılar]
      summary: Hesabın durumunu değiştirir
      description: |
        Askıya alınan (`SUSPENDED`) ya da kapatılan (`DISABLED`) hesabın bütün
        oturumları hemen sonlanır. Kişi kendi hesabının durumunu değiştiremez.
        Aktivasyonu tamamlanmamış (`PENDING`) bir hesap elle etkinleştirilemez.
        Gerekçe denetim kaydına yazılır.
      operationId: setUserStatus
      x-agora-permission: user:manage
      x-agora-mfa: true
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [status, reason]
              properties:
                status: { type: string, enum: [ACTIVE, SUSPENDED, DISABLED] }
                reason: { type: string, minLength: 1, maxLength: 500 }
      responses:
        "204":
          description: Durum değişti
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: "`FORBIDDEN` ya da `SELF_ACTION_FORBIDDEN`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "`INVALID_STATUS_CHANGE`: PENDING bir hesap elle etkinleştirilemez"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/users/{id}/activation-email:
    post:
      tags: [Kullanıcılar]
      summary: Aktivasyon e-postasını yeniden gönderir
      description: Sadece PENDING hesaplarda. Önceki bağlantılar geçersiz olur.
      operationId: resendActivationEmail
      x-agora-permission: user:manage
      x-agora-mfa: true
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "202":
          description: E-posta kuyruğa alındı
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "`ACCOUNT_NOT_PENDING`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/users/{id}/password-reset-email:
    post:
      tags: [Kullanıcılar]
      summary: Kullanıcıya parola sıfırlama e-postası gönderir
      description: Yardım masası için. Bağlantı kullanıcının e-postasına gider, yönetici görmez. Sadece ACTIVE hesaplarda.
      operationId: sendPasswordResetEmail
      x-agora-permission: user:manage
      x-agora-mfa: true
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "202":
          description: E-posta kuyruğa alındı
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "`ACCOUNT_NOT_ACTIVE`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/users/{id}/mfa/reset:
    post:
      tags: [Kullanıcılar]
      summary: Kullanıcının iki adımlı doğrulamasını sıfırlar
      description: |
        Telefonunu ve kurtarma kodlarını kaybeden kullanıcı için. Kimlik yüz yüze
        doğrulanmalıdır: bu işlem hesabı devralmanın en kısa yoludur, gerekçe zorunludur
        ve denetim izine yazılır. Kullanıcının bütün oturumları kapatılır. Yönetici kendi
        MFA'sını bu uçla kapatamaz.
      operationId: resetUserMfa
      x-agora-permission: user:manage
      x-agora-mfa: true
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason: { type: string, minLength: 1, maxLength: 500 }
      responses:
        "204":
          description: MFA kapatıldı
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: "Bkz. genel 403 kodları ve `SELF_ACTION_FORBIDDEN`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "`MFA_NOT_ENABLED`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/roles:
    get:
      tags: [Kullanıcılar]
      summary: Rol kataloğu
      description: Bütün roller, kapsam türleri ve verdikleri yetkiler. Giriş yapmış her kullanıcıya açıktır.
      operationId: listRoles
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Roller (koda göre sıralı)
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/RoleDefinition" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/users/{id}/roles:
    get:
      tags: [Kullanıcılar]
      summary: Kullanıcının rol atamaları
      description: Geçmiş, şimdiki ve ileri tarihli bütün atamalar, en yeniden eskiye.
      operationId: listUserRoles
      x-agora-permission: user:read
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Atamalar
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Assignment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Kullanıcılar]
      summary: Rol atar
      description: |
        Kapsam kimliği rolün kapsam türüne uymalıdır (üniversite geneli ve kapsamsız
        rollerde boş). Kişi kendine rol atayamaz. Aynı rol aynı kapsamda çakışan bir
        dönemde atanamaz. Değişiklik kullanıcının bir sonraki isteğinde etkilidir.
        Gerekçe zorunludur ve denetim kaydına yazılır.
      operationId: assignRole
      x-agora-permission: role:assign
      x-agora-mfa: true
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role, reason]
              properties:
                role: { type: string, examples: ["DEPARTMENT_HEAD"] }
                scope_id: { type: string, format: uuid }
                valid_from: { type: string, format: date-time, description: Boşsa hemen }
                valid_until: { type: string, format: date-time, description: Boşsa süresiz }
                reason: { type: string, minLength: 1, maxLength: 500 }
      responses:
        "201":
          description: Atama oluşturuldu
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Assignment" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: "`FORBIDDEN` ya da `SELF_ACTION_FORBIDDEN`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "`ROLE_ALREADY_ASSIGNED`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/users/{id}/roles/{assignment}/end:
    post:
      tags: [Kullanıcılar]
      summary: Rol atamasını sonlandırır
      description: |
        Başlamış atamanın bitişi "şimdi" yapılır (geçmiş korunur). Henüz başlamamış
        atama iptal edilir. Zaten bitmiş atama için bir şey yapılmaz (204). Son aktif
        sistem yöneticisinin rolü sonlandırılamaz.
      operationId: endRoleAssignment
      x-agora-permission: role:assign
      x-agora-mfa: true
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
        - name: assignment
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason: { type: string, minLength: 1, maxLength: 500 }
      responses:
        "204":
          description: Atama sonlandırıldı
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: "`FORBIDDEN` ya da `SELF_ACTION_FORBIDDEN`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "`LAST_SYSTEM_ADMIN`"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/users/{id}:
    get:
      tags: [Kullanıcılar]
      summary: Bir kullanıcının profilini döndürür
      operationId: getUser
      x-agora-permission: user:read
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Profil
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT (EdDSA)
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Cursor:
      name: cursor
      in: query
      description: Önceki yanıtın `next_cursor` değeri (opak)
      schema: { type: string }
...
  responses:
...
    BadRequest:
      description: |
        `INVALID_CLIENT` (X-Agora-Client eksik ya da geçersiz), `INVALID_BODY`
        (JSON bozuk, bilinmeyen alan, yanlış tür) ya da `VALIDATION_FAILED` (alan hataları
        `errors` dizisinde)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    UserStatus:
      type: string
      enum: [PENDING, ACTIVE, SUSPENDED, DISABLED]
...